% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_rest.R
\name{hgnc_rest}
\alias{hgnc_rest}
\title{HGNC REST Access}
\usage{
hgnc_rest()
}
\arguments{
\item{gene}{: Gene Name.}
}
\value{
A character list.
}
\description{
HGNC REST Access
}
\examples{

res <- hgnc_rest("CYBB")

}
\author{
Giordano Bruno Soares-Souza
}
