% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGeneDrug.R
\name{interactionGeneDrug}
\alias{interactionGeneDrug}
\title{Show the interactions between a given set of genes and a drugs names}
\usage{
interactionGeneDrug(
  gene = gene,
  typeReturn = "data_frame",
  interaction_sources = NULL,
  interaction_types = NULL,
  anti_neoplastic = NULL,
  gene_categories = NULL,
  clinically_actionable = NULL
)
}
\arguments{
\item{gene}{: (required) - A set of genes.}

\item{typeReturn}{: (optional) - "data_frame" (default) or "table" (first: use library(gt) and gt(res) - res as function return).}

\item{interaction_sources}{: (optional) - Can be : "DrugBank","PharmGKB","TALC","TEND","TTD"}

\item{interaction_types}{: (optional) - Can be: "activator", "inhibitor", "unknown" - (For some researches, params like interaction_types don't work properly, you should use interaction_source instead)}

\item{anti_neoplastic}{: (optional) - Can be: "true"}

\item{gene_categories}{: (optional) - Can be: "KINASE", "DNA REPAIR", "TUMOR SUPPRESSOR"}

\item{clinically_actionable}{: (optional) - Can be: "true"}
}
\value{
A data frame or gt table.
}
\description{
Show the interactions between a given set of genes and a drugs names
}
\examples{

res <- interactionGeneDrug(gene='FLT1,FLT3', interaction_sources='TALC', anti_neoplastic='true')

}
\author{
Adriano Viegas
}
