% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNCipedia_gene.R
\name{lncipedia_gene}
\alias{lncipedia_gene}
\title{Get long non-coding RNA (LNCRNAs) information}
\usage{
lncipedia_gene(genes)
}
\arguments{
\item{genes}{vector of lncRNAs symbols. Ex.: c("H19", "HOTAIR")}
}
\value{
A data frame object: chr, lncipediaTranscriptID,start, end, strand
}
\description{
This function makes it possible to search for Long non coding RNAs (LncRNAs) annotation information available in the public LNCipedia database. For this function, it is necessary to inform the gene symbol or the list of symbols of the lncRNAs to be searched. It returns a dataframe with the annotation data of the lncRNAs fetched.
}
\examples{
tt <- lncipedia_gene(c("H19", "HOTAIR", "MEG3", "SNG1", "MEG9"))
}
