% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNCipedia_search.R
\name{lncipedia_search}
\alias{lncipedia_search}
\title{Get long non-coding RNA (LNCRNAs) information}
\usage{
lncipedia_search(type, elements)
}
\arguments{
\item{type}{Specific type of information. eg: id, chromosome,start,end,high_confidence_set,keyword or class}

\item{elements}{eg: 'ENSG00000228630'}
}
\value{
A data frame object
}
\description{
This function makes it possible to search for Long non coding RNAs (LncRNAs) annotation information available in the public LNCipedia database. For this function, it is necessary to inform the type of search and the specific elements related to the type of search. It allows searching from different data entered, such as chromosome, transcript class, id, start and end of the gene. It returns a dataframe with the annotation data of the lncRNAs fetched.
}
\details{
id (optional): Any lncRNA identifier, internal or external. eg: 'ENSG00000228630'

chromosome (optional): Chromosome. eg: 'chr10'

start (optional): Start of region

end (optional): End of region

high_confidence_set (optional): Only retrieve lncRNAs in the high-confidence set. Default: false, set to 1 to enable

keyword (optional): A keyword in an abstract associated with the lncRNA. eg: 'cancer'

class (optional): Subclass. eg: 'intergenic'
}
\examples{
tt <- lncipedia_search(type = "id", elements = "ENSG00000228630")
aa <- lncipedia_search(type = "chromosome", elements = c("chr10", "chr11"))
}
