% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNCipedia_transcript.R
\name{lncipedia_transcript}
\alias{lncipedia_transcript}
\title{Get long non-coding RNA (LNCRNAs) information}
\usage{
lncipedia_transcript(transcripts)
}
\arguments{
\item{transcripts}{vector of transcripts id. Ex.: c("H19:4", "HOTAIR:1")}

\item{elements}{Gene:transcript number. eg: ('H19:4')}
}
\value{
A data frame object: refGenome, chr, TranscriptID, class, size, numExons,start, end, strand
}
\description{
This function makes it possible to search for Long non coding RNAs (LncRNAs) annotation information available in the public LNCipedia database. For this function, it is necessary to inform the symbol of the lncRNA, followed by “:” and the ID of the transcript to be searched. Ex. (“HOTAIR:1”). It is possible to fetch information from several lncRNAs by passing a vector as in the example: c(“HOTAIR:1”, “H19:3”, “H19:4”). It returns a dataframe with the annotation data of the lncRNAs fetched.
}
\examples{
tt <- lncipedia_transcript(c("H19:1", "HOTAIR:4"))

}
