% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther_map.R
\name{panther_map}
\alias{panther_map}
\title{Mapping}
\usage{
panther_map(genes, organism = "9606")
}
\arguments{
\item{genes}{Genes identifier. Maximum of 1000 Identifiers can be any of the following: Ensemble gene identifier, Ensemble protein identifier, Ensemble transcript identifier, Entrez gene id, gene symbol, NCBI GI, HGNC Id, International protein index id, NCBI UniGene id, UniProt accession andUniProt id}

\item{organism}{Taxon id. Ex.: 9606}
}
\value{
Return a dataframe

name: Pathways' name

id: Gene Ontology ID

panther_id: Annotation id from PANTHER

panther_label: Annotation label from PANTHER

gene_acession: Id of the gene from HGNC and UniProtKB
}
\description{
Maps user specified list of genes to PANTHER's genes for a specified organism and return  gene ontology annotations, PANTHER protein class annotations and Reactome annotationsof the mapped genes.
}
\examples{

data <- panther_map(genes = c("BRCA1", "VDR", "HBB"))
unique <- panther_map(genes = c("AICDA"))
}
\author{
Michele Molina e Rafaella Ferraz
}
