% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther_overrepresentation.R
\name{panther_over}
\alias{panther_over}
\title{Overrepresentation test}
\usage{
panther_over(
  genes,
  organism = "9606",
  ref_list = NULL,
  refOrganism = NULL,
  annotDataSet,
  testtype = "FISHER",
  correction = "FDR"
)
}
\arguments{
\item{genes}{Genes identifier. Maximum of 1000 Identifiers can be any of the following: Ensemble gene identifier, Ensemble protein identifier, Ensemble transcript identifier, Entrez gene id, gene symbol, NCBI GI, HGNC Id, International protein index id, NCBI UniGene id, UniProt accession andUniProt id}

\item{organism}{One taxon id required. Ex.: 9606}

\item{ref_list}{If not specified, the system will use all the genes for the specified organism. Each identifier to be delimited by comma i.e. ','. Maximum of 100000 Identifiers can be any of the following: Ensemble gene identifier, Ensemble protein identifier, Ensemble transcript identifier, Entrez gene id, gene symbol, NCBI GI, HGNC Id, International protein index id, NCBI UniGene id, UniProt accession andUniProt id}

\item{refOrganism}{This parameter is only required, if parameter 'refInputList' has been specified. Only one taxon id can be specified. Retrieve the list of supported genomes and corresponding taxon ids from the 'supportedgenomes' service.}

\item{annotDataSet}{One of the supported PANTHER annotation data types. Use the 'supportedannotdatasets' service to retrieve list of supported annotation data types. Ex: GO:0003674}

\item{testtype}{Fisher's Exact test will be used by default. Options - "FISHER" and "BINOMIAL"}

\item{correction}{correction of enrichment. Options - "FDR", "BONFERRONI" and "NONE"}
}
\value{
Return a dataframe:

number_in_list

fold_enrichment: proportion of term genes found in the input list compared to the. proportion of total term genes found in the background

fdr: correction

expected

number_in_reference

pValue

term.id: Gene Ontology ID

term.label

plus_minus
}
\description{
This test tool, compares a test gene list to a reference gene list, and determines whether a particular class (e.g. molecular function, biological process, cellular component, PANTHER protein class, the PANTHER pathway or Reactome pathway) of genes is overrepresented or underrepresented.

PANTHER tools for overrepresentation test using either Binomial or Fisher's Exact Test
}
\examples{

data <- panther_over(genes = c("BRCA1", "VDR", "HBB"), annotDataSet = "GO:0003674")
data2 <- panther_over(genes = c("BRCA1", "VDR", "HBB"), annotDataSet = "GO:0003674",ref_list = c("BRCA2", "APC"), refOrganism = "9606")
}
\author{
Michele Molina e Rafaella Ferraz
}
