% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_annot.R
\name{support_annot}
\alias{support_annot}
\title{Information about annotation data sets supported by PANTHER}
\usage{
support_annot()
}
\value{
Returns information about annotation data sets supported by PANTHER

release_date

description

id

label

version
}
\description{
Information about annotation data sets supported by PANTHER
}
\examples{
sup <- support_annot()
}
\author{
Michele Molina e Rafaella Ferraz
}
