% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_convert.R
\name{ensembl_convert}
\alias{ensembl_convert}
\title{Convert the co-ordinates of one assembly to another}
\usage{
ensembl_convert(asm_one, asm_two, chr, coord_start, coord_end, strand)
}
\arguments{
\item{asm_one}{Version of the input assembly    -    (Ex: GRCh37)}

\item{asm_two}{Version of the output assembly    -    (Ex: GRCh38)}

\item{chr}{type cromossomo    X}

\item{coord_start}{Coordinate start}

\item{coord_end}{Coordinate End}

\item{strand}{DNA strand}
}
\value{
A data frame.
# assembly seq_region_name     end coord_system   start strand
# 1   GRCh37               X 1000100   chromosome 1000000      1
# 2   GRCh38               X 1039365   chromosome 1039265      1
}
\description{
Convert the co-ordinates of one assembly to another
}
\examples{

fg <- ensembl_convert(asm_one ="GRCh37",
chr = "x",
coord_start ="1000000",
coord_end ="1000100",
asm_two ="GRCH38",
strand = "1")

# /map/human/GRCh37/X:1000000..1000100:1/GRCh38?content-type=application/json





}
\author{
Eric Baia
}
