% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionDrugGene.R
\name{interactionDrugGene}
\alias{interactionDrugGene}
\title{Show the interactions between a given set of drugs and a gene names}
\usage{
interactionDrugGene(drugs, itype)
}
\arguments{
\item{drugs}{: A set of drugs (required)*.}

\item{itype}{: Interaction type - can be: inhibitor or activator (optional)*.}
}
\value{
A character list.
}
\description{
Show the interactions between a given set of drugs and a gene names
}
\examples{

res <- interactionDrugGene("VATALANIB", "inhibitor")
res
VATALANIB - KDR
VATALANIB - CSF1R
}
\author{
Adriano Viegas
}
