% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGeneDrug.R
\name{interactionGeneDrug}
\alias{interactionGeneDrug}
\title{Show the interactions between a given set of gene and a drug names}
\usage{
interactionGeneDrug(genes, itype)
}
\arguments{
\item{genes}{: A set of genes (required)*.}

\item{itype}{: Interaction type - can be: inhibitor or activator (optional)*.}
}
\value{
A character list.
}
\description{
Show the interactions between a given set of gene and a drug names
}
\examples{

res <- interactionGeneDrug("FLT1", "inhibitor")
res
FLT1 - DOVITINIB
FLT1 - LINIFANIB

}
\author{
Adriano Viegas
}
